replaced departments for collapsible to create reusable to use in faq
